Fat refactor of Tokens section
