+++ src/components/v2v3/V2V3Project/V2V3ManageTokensSection/AccountBalanceDescription/V2RedeemModal.tsx
@@ -1,15 +1,9 @@
 import { t, Trans } from '@lingui/macro'
 import { Descriptions, Form, Space } from 'antd'
 import { useForm } from 'antd/lib/form/Form'
-import { useWallet } from 'hooks/Wallet'
-
-import InputAccessoryButton from 'components/InputAccessoryButton'
-
-import { useContext, useState } from 'react'
-import { formatWad, fromWad, parseWad } from 'utils/format/formatNumber'
-
 import Callout from 'components/Callout'
 import ETHAmount from 'components/currency/ETHAmount'
+import InputAccessoryButton from 'components/InputAccessoryButton'
 import FormattedNumberInput from 'components/inputs/FormattedNumberInput'
 import { MemoFormInput } from 'components/Project/PayProjectForm/MemoFormInput'
 import { RedeemAMMPrices } from 'components/Project/RedeemAMMPrices'
@@ -19,6 +13,9 @@ import { V2V3ProjectContext } from 'contexts/v2v3/V2V3ProjectContext'
 import { useETHReceivedFromTokens } from 'hooks/v2v3/contractReader/ETHReceivedFromTokens'
 import useTotalBalanceOf from 'hooks/v2v3/contractReader/TotalBalanceOf'
 import { useRedeemTokensTx } from 'hooks/v2v3/transactor/RedeemTokensTx'
+import { useWallet } from 'hooks/Wallet'
+import { useContext, useState } from 'react'
+import { formatWad, fromWad, parseWad } from 'utils/format/formatNumber'
 import { emitErrorNotification } from 'utils/notifications'
 import { tokenSymbolText } from 'utils/tokenSymbolText'
 import { V2V3_CURRENCY_USD } from 'utils/v2v3/currency'
