+++ src/components/v2v3/V2V3Project/V2V3ManageTokensSection/AccountBalanceDescription/index.ts
@@ -1,0 +1,1 @@
+export { AccountBalanceDescription } from './AccountBalanceDescription'
