--- src/components/v2v3/V2V3Project/V2V3ManageTokensSection/V1ProjectTokensSection/index.tsx
@@ -1,1 +1,0 @@
-export * from './V1ProjectTokensSection'
